requirements.txt
requests==2.31.0
mysql-connector-python==8.2.0
python-dateutil==2.8.2
click==8.1.7
python-dotenv==1.0.0
